separate package for competitors
