+++ src/Lesson_1/Marafon/Competitors/Competitor.java
@@ -1,4 +1,4 @@
-package Lesson_1.Marafon;
+package Lesson_1.Marafon.Competitors;
 
 public interface Competitor {
     void run(int dist);
