+++ src/Lesson_1/Marafon/Cross.java
@@ -1,5 +1,7 @@
 package Lesson_1.Marafon;
 
+import Lesson_1.Marafon.Competitors.Competitor;
+
 public class Cross extends Obstacle {
     int length;
