+++ src/Lesson_1/Marafon/Main.java
@@ -1,5 +1,10 @@
 package Lesson_1.Marafon;
 
+import Lesson_1.Marafon.Competitors.Cat;
+import Lesson_1.Marafon.Competitors.Competitor;
+import Lesson_1.Marafon.Competitors.Dog;
+import Lesson_1.Marafon.Competitors.Human;
+
 public class Main {
     public static void main(String[] args) {
         Competitor[] competitors = {new Human("Боб"), new Cat("Барсик"), new Dog("Бобик")};
