+++ src/Lesson_1/Marafon/Obstacle.java
@@ -1,5 +1,7 @@
 package Lesson_1.Marafon;
 
+import Lesson_1.Marafon.Competitors.Competitor;
+
 public abstract class Obstacle {
     public abstract void doIt(Competitor competitor);
 }
