+++ src/Lesson_1/Marafon/Wall.java
@@ -1,5 +1,7 @@
 package Lesson_1.Marafon;
 
+import Lesson_1.Marafon.Competitors.Competitor;
+
 public class Wall extends Obstacle {
     int height;
